Add Maximum line to the burndown chart
The ideal burndown line is necessary for teams to understand how
much they need to burn down per day to hit their target, but if
you're above the line, it can be difficult sometimes to tell whether
the overage can be consumed by the team. Enter the Maximum line,
which takes the team's initial capacity and burns it down over the
life of the sprint. As long as the team is under the max line, they
can reasonably consume an overage. Going over the line is an indicator
that the sprint likely needs to change or it will fail.